+++ vortex-file/src/file/read/layouts/flat.rs
@@ -5,6 +5,8 @@ use bytes::Bytes;
 use vortex_array::{Array, Context};
 use vortex_error::{vortex_bail, VortexResult};
 use vortex_flatbuffers::footer;
+use vortex_ipc::messages::reader::ArrayMessageReader;
+use vortex_ipc::stream_writer::ByteRange;
 
 use crate::file::read::cache::RelativeLayoutCache;
 use crate::file::read::mask::RowMask;
@@ -12,8 +14,6 @@ use crate::file::{
     BatchRead, LayoutDeserializer, LayoutId, LayoutReader, LayoutSpec, MessageLocator, Scan,
     FLAT_LAYOUT_ID,
 };
-use crate::messages::reader::ArrayMessageReader;
-use crate::stream_writer::ByteRange;
 
 #[derive(Debug)]
 pub struct FlatLayoutSpec;
@@ -127,13 +127,13 @@ mod tests {
     use vortex_array::{Context, IntoArray, IntoArrayVariant};
     use vortex_dtype::PType;
     use vortex_expr::{BinaryExpr, Identity, Literal, Operator};
+    use vortex_ipc::messages::writer::MessageWriter;
+    use vortex_ipc::stream_writer::ByteRange;
 
     use crate::file::read::cache::{LazilyDeserializedDType, RelativeLayoutCache};
     use crate::file::read::layouts::flat::FlatLayout;
     use crate::file::read::layouts::test_read::{filter_read_layout, read_layout};
     use crate::file::{LayoutMessageCache, RowFilter, Scan};
-    use crate::messages::writer::MessageWriter;
-    use crate::stream_writer::ByteRange;
 
     async fn read_only_layout(
         cache: Arc<RwLock<LayoutMessageCache>>,
