+++ vortex-file/src/file/read/mod.rs
@@ -25,11 +25,11 @@ pub use filtering::RowFilter;
 pub use recordbatchreader::{AsyncRuntime, VortexRecordBatchReader};
 pub use stream::VortexFileArrayStream;
 use vortex_expr::ExprRef;
+use vortex_ipc::stream_writer::ByteRange;
 pub use vortex_schema::projection::Projection;
 pub use vortex_schema::Schema;
 
 pub use crate::file::read::mask::RowMask;
-use crate::stream_writer::ByteRange;
 
 // Recommended read-size according to the AWS performance guide
 pub const INITIAL_READ_SIZE: usize = 8 * 1024 * 1024;
