+++ vortex-file/src/file/write/layout.rs
@@ -1,11 +1,11 @@
 use bytes::Bytes;
 use flatbuffers::{FlatBufferBuilder, WIPOffset};
 use vortex_flatbuffers::{footer as fb, WriteFlatBuffer};
+use vortex_ipc::stream_writer::ByteRange;
 
 use crate::file::{
     LayoutId, CHUNKED_LAYOUT_ID, COLUMNAR_LAYOUT_ID, FLAT_LAYOUT_ID, INLINE_SCHEMA_LAYOUT_ID,
 };
-use crate::stream_writer::ByteRange;
 
 #[derive(Debug, Clone)]
 pub struct Layout {
