+++ vortex/src/lib.rs
@@ -4,8 +4,8 @@ pub use {
     vortex_datetime_dtype as datetime_dtype, vortex_datetime_parts as datetime_parts,
     vortex_dict as dict, vortex_dtype as dtype, vortex_error as error, vortex_expr as expr,
     vortex_fastlanes as fastlanes, vortex_file as file, vortex_flatbuffers as flatbuffers,
-    vortex_fsst as fsst, vortex_io as io, vortex_proto as proto, vortex_roaring as roaring,
-    vortex_runend as runend, vortex_runend_bool as runend_bool,
+    vortex_fsst as fsst, vortex_io as io, vortex_ipc as ipc, vortex_proto as proto,
+    vortex_roaring as roaring, vortex_runend as runend, vortex_runend_bool as runend_bool,
     vortex_sampling_compressor as sampling_compressor, vortex_scalar as scalar,
     vortex_schema as schema, vortex_zigzag as zigzag,
 };
